Add specs for importer, dir, and writer

+++ plugins/gatsby-ample-importer/utils/dir.js
@@ -16,6 +16,7 @@ exports.initDir = dir => {
  * @param {string} dir Path to directory
  */
 exports.cleanDir = dir => {
+  if (!fs.existsSync(dir)) return
   const existingFiles = fs.readdirSync(dir)
   existingFiles.map(file => fs.unlinkSync(path.join(dir, file)))
 }
